+++ smartcontract/sdk/rs/src/commands/tenant/add_administrator.rs
@@ -16,7 +16,7 @@ impl AddAdministratorTenantCommand {
             .execute(client)
             .map_err(|_err| eyre::eyre!("Globalstate not initialized"))?;
 
-        client.execute_transaction(
+        client.execute_authorized_transaction(
             DoubleZeroInstruction::TenantAddAdministrator(TenantAddAdministratorArgs {
                 administrator: self.administrator,
             }),